refactor: 불필요한 파일 삭제, 구조 변경 및 PostsManagerPage 경로 수정
