--- src/entities/comment/model/index.ts
@@ -1,1 +1,0 @@
-// 댓글 엔티티 타입
