--- src/entities/comment/ui/CommentItem/index.tsx
@@ -1,1 +1,0 @@
-// 댓글 아이템 컴포넌트
