--- src/entities/post/model/index.ts
@@ -1,1 +1,0 @@
-// 게시물 엔티티 타입 및 유틸
