--- src/entities/post/ui/PostActions/index.tsx
@@ -1,1 +1,0 @@
-// 게시물 액션 버튼
