--- src/entities/post/ui/PostCard/index.tsx
@@ -1,1 +1,0 @@
-// 게시물 카드 컴포넌트
