--- src/entities/user/model/index.ts
@@ -1,1 +1,0 @@
-// 사용자 엔티티 타입
