--- src/entities/user/ui/UserAvatar/index.tsx
@@ -1,1 +1,0 @@
-// 사용자 아바타 컴포넌트
