--- src/features/comments/api/index.ts
@@ -1,1 +1,0 @@
-// 댓글 API 통신
