--- src/features/comments/model/index.ts
@@ -1,1 +1,0 @@
-// 댓글 관련 Zustand 스토어
