--- src/features/comments/ui/CommentForm/index.tsx
@@ -1,1 +1,0 @@
-// 댓글 추가/수정 폼
