--- src/features/comments/ui/CommentList/index.tsx
@@ -1,1 +1,0 @@
-// 댓글 목록 컴포넌트
