--- src/features/posts/api/index.ts
@@ -1,1 +1,0 @@
-// 게시물 API 통신
