--- src/features/posts/model/index.ts
@@ -1,1 +1,0 @@
-// 게시물 관련 Zustand 스토어
