--- src/features/posts/ui/PostDetail/index.tsx
@@ -1,1 +1,0 @@
-// 게시물 상세 다이얼로그
