--- src/features/posts/ui/PostForm/index.tsx
@@ -1,1 +1,0 @@
-// 게시물 추가/수정 폼
