--- src/features/posts/ui/index.ts
@@ -1,1 +1,0 @@
-export { default as PostsTable } from "./PostsTable"
