--- src/features/users/model/index.ts
@@ -1,1 +1,0 @@
-// 사용자 관련 Zustand 스토어
