--- src/features/users/ui/UserModal/index.tsx
@@ -1,1 +1,0 @@
-// 사용자 정보 모달
